Updated the docker env variables and paths for the road accidents csv data ingestion dag. Updated readme with the Volumes dir structure and added some Volume directories to not be ignored by git (but they files yes). The whole application runs including mlflow
